Remove consul-template support. Do not redownload.
Skip download if file we are trying to download already exists on the
disk. We are anyway not deleting SOURCES folder when recreating build
tree.

Remove explicit consul template support as the script we are trying to
make should be generic enough to easily setup packaging for Go-based
projects. Consul template can be easily built by changing the package
name and download root URL at the top of the file. It will make more
sense to expose them as command line arguments/switches.